copyfilesync for .lcok protection

+++ src/actions/install.ts
@@ -1,7 +1,7 @@
 import ActionInterface from '../ActionInterface'
 import GeneratorInterface from '../GeneratorInterface'
 import {composer} from '../helpers'
-import {copyFileSync} from 'fs'
+import {copyFileSync, existsSync} from 'fs'
 
 export default class InstallAction implements ActionInterface {
   /**
@@ -23,7 +23,9 @@ export default class InstallAction implements ActionInterface {
    */
   exec(generator: GeneratorInterface): void {
     const packages = generator.getComposerRequirements()
-    copyFileSync('composer.lock', '.composer.lock.bkp')
+    if (existsSync('composer.lock')) {
+      copyFileSync('composer.lock', '.composer.lock.bkp')
+    }
     copyFileSync('composer.json', '.composer.json.bkp')
     composer(
       ['require', '--dev', '--no-progress', '--no-suggest'].concat(packages)
